Refactor HotRodConfiguration into wildfly-clustering-ee-hotrod.
