+++ clustering/ee/hotrod/src/main/java/org/wildfly/clustering/ee/hotrod/HotRodConfiguration.java
@@ -20,20 +20,30 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.clustering.web.hotrod.session;
+package org.wildfly.clustering.ee.hotrod;
 
 import org.infinispan.client.hotrod.RemoteCache;
+import org.wildfly.clustering.ee.Batcher;
+import org.wildfly.clustering.ee.cache.CacheConfiguration;
 import org.wildfly.clustering.ee.cache.CacheProperties;
-import org.wildfly.clustering.ee.hotrod.RemoteCacheProperties;
+import org.wildfly.clustering.ee.cache.tx.TransactionBatch;
+import org.wildfly.clustering.ee.hotrod.tx.HotRodBatcher;
 
 /**
  * @author Paul Ferraro
  */
-public interface HotRodSessionMetaDataFactoryConfiguration {
+public interface HotRodConfiguration extends CacheConfiguration {
 
+    @Override
     <CK, CV> RemoteCache<CK, CV> getCache();
 
+    @Override
     default CacheProperties getCacheProperties() {
         return new RemoteCacheProperties(this.getCache());
+    }
+
+    @Override
+    default Batcher<TransactionBatch> getBatcher() {
+        return new HotRodBatcher(this.getCache());
     }
 }
