+++ clustering/web/hotrod/src/main/java/org/wildfly/clustering/web/hotrod/session/HotRodSessionAttributesFactoryConfiguration.java
@@ -23,6 +23,7 @@
 package org.wildfly.clustering.web.hotrod.session;
 
 import org.wildfly.clustering.ee.cache.CacheProperties;
+import org.wildfly.clustering.ee.hotrod.HotRodConfiguration;
 import org.wildfly.clustering.web.cache.session.SessionAttributesFactoryConfiguration;
 
 /**
@@ -33,10 +34,10 @@ import org.wildfly.clustering.web.cache.session.SessionAttributesFactoryConfigur
  * @param <V> attributes cache entry type
  * @param <SV> attributes serialized form type
  */
-public interface HotRodSessionAttributesFactoryConfiguration<S, C, L, V, SV> extends SessionAttributesFactoryConfiguration<S, C, L, V, SV>, HotRodSessionMetaDataFactoryConfiguration {
+public interface HotRodSessionAttributesFactoryConfiguration<S, C, L, V, SV> extends SessionAttributesFactoryConfiguration<S, C, L, V, SV>, HotRodConfiguration {
 
     @Override
     default CacheProperties getCacheProperties() {
-        return HotRodSessionMetaDataFactoryConfiguration.super.getCacheProperties();
+        return HotRodConfiguration.super.getCacheProperties();
     }
 }
